+++ Sources/ChangeLoggerCLI/CommandLineError.swift
@@ -4,6 +4,8 @@ enum CommandLineError: Error {
   case unkownCommand(String)
   case missingChangeLogYAML
   case missingCommitYAML
+  case missingReleaseSummary
+  case missingVersion
 
   var description: String {
     switch self {
@@ -13,6 +15,10 @@ enum CommandLineError: Error {
         return "must supply file path"
       case .missingCommitYAML:
         return "must supply file path"
+      case .missingReleaseSummary:
+        return "must supply summary for release"
+      case .missingVersion:
+        return "Version must be supplied"
     }
   }
 }