+++ Sources/ChangeLoggerCLI/CommandLineInterface.swift
@@ -12,6 +12,7 @@ public class ChangeLoggerCommandLine {
   public init() {
     registry.register(command: InitCommand.self)
     registry.register(command: LogCommand.self)
+    registry.register(command: ReleaseCommand.self)
     registry.register(command: MarkdownCommand.self)
   }